refactor(RoomView): move navigateToScreen into services
