+++ app/views/RoomView/services/__tests__/closeLivechat.test.ts
@@ -2,7 +2,7 @@ import { closeLivechat as closeLivechatService } from '../../../../lib/methods/h
 import { showErrorAlert } from '../../../../lib/methods/helpers/info';
 import log from '../../../../lib/methods/helpers/log';
 import { getDepartmentInfo, getTagsList } from '../../../../lib/services/restApi';
-import { navigateToScreen, type TRoomStackNavigation } from '../../hooks/navigateToScreen';
+import { navigateToScreen, type TRoomStackNavigation } from '../navigateToScreen';
 import { closeLivechat } from '../closeLivechat';
 
 jest.mock('../../../../lib/methods/helpers/closeLivechat', () => ({ closeLivechat: jest.fn(() => Promise.resolve()) }));
@@ -10,7 +10,7 @@ jest.mock('../../../../lib/services/restApi', () => ({
 	getDepartmentInfo: jest.fn(),
 	getTagsList: jest.fn()
 }));
-jest.mock('../../hooks/navigateToScreen', () => ({ navigateToScreen: jest.fn() }));
+jest.mock('../navigateToScreen', () => ({ navigateToScreen: jest.fn() }));
 jest.mock('../../../../lib/methods/helpers/info', () => ({ showErrorAlert: jest.fn() }));
 jest.mock('../../../../lib/methods/helpers/log', () => ({ __esModule: true, default: jest.fn() }));
